requirements.txt
youtube-dl
deemix
python-dotenv
discord.py
art
pydrive